add analyze image

+++ image-ai-backend/src/main/java/com/ftnam/image_ai_backend/dto/request/HistoryRequest.java
@@ -23,6 +23,6 @@ public class HistoryRequest {
     private String result;
 
     @NotNull
-    private float confident;
+    private double confident;
 
 }
