+++ image-ai-backend/src/main/java/com/ftnam/image_ai_backend/entity/History.java
@@ -28,7 +28,7 @@ public class History {
 
     private String result;
 
-    private float confident;
+    private double confident;
 
     @Column(name = "created_at")
     @CreationTimestamp
